moved new types to different namespaces per code review comments
