--- Fabric.Identity.API/Models/CustomGrantType.cs
@@ -1,7 +1,0 @@
-﻿namespace Fabric.Identity.API.Models
-{
-    public static class CustomGrantType
-    {
-        public const string Delegation = "delegation";
-    }
-}
